refactor(chalice): moved favorite sessions to CH
