+++ api/chalicelib/core/sessions/sessions_favorite/sessions_favorite.py
@@ -1,5 +1,9 @@
+import logging
+
 import schemas
-from chalicelib.utils import pg_client
+from chalicelib.utils import pg_client, ch_client, exp_ch_helper
+
+logger = logging.getLogger(__name__)
 
 
 def add_favorite_session(context: schemas.CurrentContext, project_id, session_id):
@@ -13,6 +17,7 @@ def add_favorite_session(context: schemas.CurrentContext, project_id, session_id
         )
         row = cur.fetchone()
     if row:
+        add_favorite_session_to_ch(project_id=project_id, user_id=context.user_id, session_id=session_id)
         return {"data": {"sessionId": session_id}}
     return {"errors": ["something went wrong"]}
 
@@ -29,6 +34,7 @@ def remove_favorite_session(context: schemas.CurrentContext, project_id, session
         )
         row = cur.fetchone()
     if row:
+        remove_favorite_session_from_ch(project_id=project_id, user_id=context.user_id, session_id=session_id)
         return {"data": {"sessionId": session_id}}
     return {"errors": ["something went wrong"]}
 
@@ -60,12 +66,29 @@ def get_start_end_timestamp(project_id, user_id):
     with pg_client.PostgresClient() as cur:
         cur.execute(
             cur.mogrify(
-                """SELECT max(start_ts) AS max_start_ts, min(start_ts) AS min_start_ts                                                
-                    FROM public.user_favorite_sessions INNER JOIN sessions USING(session_id)
-                    WHERE
-                     user_favorite_sessions.user_id = %(userId)s
+                """SELECT max(start_ts) AS max_start_ts, min(start_ts) AS min_start_ts
+                   FROM public.user_favorite_sessions
+                            INNER JOIN sessions USING (session_id)
+                   WHERE user_favorite_sessions.user_id = %(userId)s
                      AND project_id = %(project_id)s;""",
                 {"userId": user_id, "project_id": project_id})
         )
         r = cur.fetchone()
     return (0, 0) if r is None else (r["min_start_ts"], r["max_start_ts"])
+
+
+def add_favorite_session_to_ch(project_id, user_id, session_id, sign=1):
+    try:
+        with ch_client.ClickHouseClient() as cur:
+            query = f"""INSERT INTO {exp_ch_helper.get_user_favorite_sessions_table()}(project_id,user_id, session_id, sign) 
+                        VALUES (%(project_id)s,%(userId)s,%(sessionId)s,%(sign)s);"""
+            params = {"userId": user_id, "sessionId": session_id, "project_id": project_id, "sign": sign}
+            cur.execute(query=query, parameters=params)
+
+    except Exception as err:
+        logger.error("------- Exception while adding favorite session to CH")
+        logger.error(err)
+
+
+def remove_favorite_session_from_ch(project_id, user_id, session_id):
+    add_favorite_session_to_ch(project_id=project_id, user_id=user_id, session_id=session_id, sign=-1)
