+++ ee/api/chalicelib/core/sessions/sessions_favorite/sessions_favorite_ee.py
@@ -4,29 +4,10 @@ from decouple import config
 
 import schemas
 from chalicelib.core.sessions import sessions_mobs, sessions_devtool
-from .sessions_favorite import add_favorite_session as _add_favorite_session, \
-    remove_favorite_session as _remove_favorite_session, \
-    favorite_session_exists
-from chalicelib.utils import ch_client, exp_ch_helper
 from chalicelib.utils.storage import extra
+from .sessions_favorite import add_favorite_session, remove_favorite_session, favorite_session_exists
 
 logger = logging.getLogger(__name__)
-
-
-def add_favorite_session(context: schemas.CurrentContext, project_id, session_id):
-    result = _add_favorite_session(context=context, project_id=project_id, session_id=session_id)
-    if "data" in result:
-        add_favorite_session_to_ch(project_id=project_id, user_id=context.user_id,
-                                   session_id=session_id)
-    return result
-
-
-def remove_favorite_session(context: schemas.CurrentContext, project_id, session_id):
-    result = _remove_favorite_session(context=context, project_id=project_id, session_id=session_id)
-    if "data" in result:
-        remove_favorite_session_from_ch(project_id=project_id, user_id=context.user_id,
-                                        session_id=session_id)
-    return result
 
 
 def favorite_session(context: schemas.CurrentContext, project_id, session_id):
@@ -56,20 +37,3 @@ def favorite_session(context: schemas.CurrentContext, project_id, session_id):
             print(str(e))
 
     return add_favorite_session(context=context, project_id=project_id, session_id=session_id)
-
-
-def add_favorite_session_to_ch(project_id, user_id, session_id, sign=1):
-    try:
-        with ch_client.ClickHouseClient() as cur:
-            query = f"""INSERT INTO {exp_ch_helper.get_user_favorite_sessions_table()}(project_id,user_id, session_id, sign) 
-                        VALUES (%(project_id)s,%(userId)s,%(sessionId)s,%(sign)s);"""
-            params = {"userId": user_id, "sessionId": session_id, "project_id": project_id, "sign": sign}
-            cur.execute(query=query, parameters=params)
-
-    except Exception as err:
-        logger.error("------- Exception while adding favorite session to CH")
-        logger.error(err)
-
-
-def remove_favorite_session_from_ch(project_id, user_id, session_id):
-    add_favorite_session_to_ch(project_id=project_id, user_id=user_id, session_id=session_id, sign=-1)
