+++ ee/api/clean-dev.sh
@@ -33,6 +33,8 @@ rm -rf ./chalicelib/core/mobile.py
 rm -rf ./chalicelib/core/saved_search.py
 rm -rf ./chalicelib/core/sessions/*.py
 rm -rf ./chalicelib/core/sessions/sessions_viewed
+rm -rf ./chalicelib/core/sessions/sessions_favorite/sessions_favorite.py
+rm -rf ./chalicelib/core/sessions/sessions_devtool/sessions_devtool.py
 rm -rf ./chalicelib/core/metrics/modules
 rm -rf ./chalicelib/core/socket_ios.py
 rm -rf ./chalicelib/core/sourcemaps
